Reorganize form components into v8/v9 folders
